add some tests for imperative message
